third_party/WebKit/LayoutTests/media/video-muted.html

<!DOCTYPE html>
<title>Test "muted" attribute.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    assert_false(video.muted);
    video.muted = true;
    assert_true(video.muted);

    video.oncanplaythrough = t.step_func_done(function () {
        assert_true(video.muted);
        video.muted = false;
        assert_false(video.muted);
    });

    video.src = "content/test.ogv";
});
</script>
